+++ src/components/MlMeasureTool/MlMeasureTool.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import MlFeatureEditor from "../MlFeatureEditor/MlFeatureEditor";
 import * as turf from "@turf/turf";
 
@@ -6,28 +6,55 @@ interface MlMeasureToolProps {
   /**
    * String that specify if the Tool measures an area ("polygon") or length ("line")
    */
-  measureType: string;
+  measureType?: string;
+  /**
+   * String that dictates which unit of measurement is used
+   */
+  unit?: turf.Units;
+}
+
+//const unitSquareConvert = {
+//  kilometers: 1,
+//  miles: 1 / 2.58998811,
+//};
+function getUnitSquareMultiplier(measureType:string | undefined) {
+  return measureType === "miles" ? 1 / 2.58998811 : 1;
+}
+function getUnitLabel(measureType:string | undefined) {
+  return measureType === "miles" ? 'mi' : 'km';
 }
 
 const MlMeasureTool = (props: MlMeasureToolProps) => {
-  const [length, setLength] = useState(0)
+  const [length, setLength] = useState(0);
+  const [currentFeatures, setCurrentFeatures] = useState([undefined]);
+
+  useEffect(() => {
+    if (currentFeatures[0]) {
+      setLength(
+        props.measureType === "polygon"
+          ? (turf.area(currentFeatures[0]) / 1000000) * getUnitSquareMultiplier(props.unit)
+          : turf.length(currentFeatures[0], { units: props.unit })
+      );
+    }
+  }, [props.unit, currentFeatures]);
 
   return (
     <>
       <MlFeatureEditor
-        onChange={(features) => {
-          console.log(features);
-          if(features[0]) {
-            setLength(props.measureType === "polygon" ? turf.area(features[0]) / 1000000 : turf.length(features[0]));
-          }}}
-        mode = {props.measureType === "polygon" ? "custom_polygon" : "draw_line_string"}
+        onChange={(features:any) => {
+          setCurrentFeatures(features);
+        }}
+        mode={props.measureType === "polygon" ? "custom_polygon" : "draw_line_string"}
       />
-      {props.measureType === "polygon" ? "Area" : "Length"}: {length.toFixed(2)} {props.measureType === "polygon" ? "km²" : "km"}
-  </>);
+      {length.toFixed(2)} {getUnitLabel(props.unit)}
+      {props.measureType === "polygon" ? "²" : ""}
+    </>
+  );
 };
 
 MlMeasureTool.defaultProps = {
   mapId: undefined,
   measureType: "line",
+  unit: "kilometers",
 };
 export default MlMeasureTool;