+++ src/components/MlNavigationTools/MlNavigationTools.tsx
@@ -3,18 +3,52 @@ import Button from "@mui/material/Button";
 import ButtonGroup from "@mui/material/ButtonGroup";
 import ControlPointIcon from "@mui/icons-material/ControlPoint";
 import RemoveCircleOutlineIcon from "@mui/icons-material/RemoveCircleOutline";
+import Box from "@mui/material/Box";
+import { SxProps } from "@mui/material";
 
 import MlNavigationCompass from "../MlNavigationCompass/MlNavigationCompass";
 import MlFollowGps from "../MlFollowGps/MlFollowGps";
 import useMediaQuery from "@mui/material/useMediaQuery";
 import useMap from "../../hooks/useMap";
+import MlCenterPosition from "../MlCenterPosition/MlCenterPosition";
 
 interface MlNavigationToolsProps {
+  /**
+   * Id of the target MapLibre instance in mapContext
+   */
   mapId?: string;
+  /**
+   * The layerId of an existing layer this layer should be rendered visually beneath
+   * https://maplibre.org/maplibre-gl-js-docs/api/map/#map#addlayer - see "beforeId" property
+   */
   insertBeforeLayer?: string;
+  /**
+   * Show 3D button
+   */
+  show3DButton?: boolean;
+  /**
+   * Show zoom button
+   */
+  showZoomButtons?: boolean;
+  /**
+   * Show follow GPS button
+   */
+  showFollowGpsButton?: boolean;
+  /**
+   * Show center on current position button
+   */
+  showCenterLocationButton?: boolean;
+  /**
+   * Additional JSX Elements to be rendered below MlNavigationTools buttons
+   */
+  children?: JSX.Element;
+  /**
+   * Style attribute for NavigationTools container
+   */
+  sx?: SxProps;
 }
 
- /**
+/**
  * @component
  */
 const MlNavigationTools = (props: MlNavigationToolsProps) => {
@@ -35,31 +69,35 @@ const MlNavigationTools = (props: MlNavigationToolsProps) => {
     //border: "1px solid #bbb",
     //boxShadow: "0px 0px 4px rgba(0,0,0,.5)",
     margin: 0.15,
-    fontSize: mediaIsMobile ? "1.5em" : "1.2em",
+    fontSize: mediaIsMobile ? "1.4em" : "1.2em",
     ":hover": {
       backgroundColor: "#515151",
     },
     color: "#ececec",
   };
+  const iconStyle = {
+    fontSize: buttonStyle.fontSize,
+  };
 
   useEffect(() => {
     if (!mapHook.map) return;
 
-    mapHook.map.on("pitchend", () => {
-      if (!mapHook.map) return;
+    mapHook.map.on(
+      "pitchend",
+      () => {
+        if (!mapHook.map) return;
 
-      setPitch(mapHook.map.map.getPitch());
-    },mapHook.componentId);
+        setPitch(mapHook.map.map.getPitch());
+      },
+      mapHook.componentId
+    );
     setPitch(mapHook.map.map.getPitch());
   }, [mapHook.map, props.mapId]);
 
   const zoomIn = () => {
     if (!mapHook.map) return;
 
-    if (
-      mapHook.map.map.transform._zoom + 0.5 <=
-      mapHook.map.map.transform._maxZoom
-    ) {
+    if (mapHook.map.map.transform._zoom + 0.5 <= mapHook.map.map.transform._maxZoom) {
       mapHook.map.map.easeTo({ zoom: mapHook.map.map.transform._zoom + 0.5 });
     }
   };
@@ -67,10 +105,7 @@ const MlNavigationTools = (props: MlNavigationToolsProps) => {
   const zoomOut = () => {
     if (!mapHook.map) return;
 
-    if (
-      mapHook.map.map.transform._zoom - 0.5 >=
-      mapHook.map.map.transform._minZoom
-    ) {
+    if (mapHook.map.map.transform._zoom - 0.5 >= mapHook.map.map.transform._minZoom) {
       mapHook.map.map.easeTo({ zoom: mapHook.map.map.transform._zoom - 0.5 });
     }
   };
@@ -86,14 +121,14 @@ const MlNavigationTools = (props: MlNavigationToolsProps) => {
   };
 
   return (
-    <div
-      style={{
+    <Box
+      sx={{
         zIndex: 501,
         position: "absolute",
-        right: mediaIsMobile ? "15px" : "5px",
-        bottom: mediaIsMobile ? "40px" : "20px",
         display: "flex",
         flexDirection: "column",
+        ...(mediaIsMobile ? { margin: "20px 10px 20px 10px" } : {}),
+        ...props.sx,
       }}
     >
       <MlNavigationCompass
@@ -108,10 +143,20 @@ const MlNavigationTools = (props: MlNavigationToolsProps) => {
           boxShadow: "0px 0px 18px rgba(0,0,0,.5)",
         }}
       />
-      <Button sx={{ ...buttonStyle, fontSize:mediaIsMobile?'1.4em':'1em', fontWeight: 600 }} onClick={adjustPitch}>
-        {pitch ? "2D" : "3D"}
-      </Button>
-      <MlFollowGps style={{ ...(({ color, ...rest }) => rest)(buttonStyle) }} />
+      {props.show3DButton && (
+        <Button
+          sx={{ ...buttonStyle, fontSize: mediaIsMobile ? "1.4em" : "1em", fontWeight: 600 }}
+          onClick={adjustPitch}
+        >
+          {pitch ? "2D" : "3D"}
+        </Button>
+      )}
+      {props.showFollowGpsButton && (
+        <MlFollowGps buttonSx={{ ...(({ color, ...rest }) => rest)(buttonStyle) }} />
+      )}
+      {props.showCenterLocationButton && (
+        <MlCenterPosition style={{ ...(({ color, ...rest }) => rest)(buttonStyle) }} />
+      )}
       <ButtonGroup
         orientation="vertical"
         sx={{
@@ -121,19 +166,35 @@ const MlNavigationTools = (props: MlNavigationToolsProps) => {
           "Button:hover": { border: "none" },
         }}
       >
-        <Button sx={{ ...buttonStyle, color: "#ececec" }} onClick={zoomIn}>
-          <ControlPointIcon
-            sx={{ fontSize: mediaIsMobile ? "1.5em" : "1.2em" }}
-          />
-        </Button>
-        <Button sx={{ ...buttonStyle, color: "#ececec" }} onClick={zoomOut}>
-          <RemoveCircleOutlineIcon
-            sx={{ fontSize: mediaIsMobile ? "1.5em" : "1.2em" }}
-          />
-        </Button>
+        {props.showZoomButtons && (
+          <>
+            <Button sx={{ ...buttonStyle }} onClick={zoomIn}>
+              <ControlPointIcon sx={{ ...iconStyle }} />
+            </Button>
+            <Button sx={{ ...buttonStyle }} onClick={zoomOut}>
+              <RemoveCircleOutlineIcon sx={{ ...iconStyle }} />
+            </Button>
+          </>
+        )}
       </ButtonGroup>
-    </div>
+      {props.children &&
+        React.cloneElement(props.children, {
+          sx: { ...buttonStyle },
+        })}
+    </Box>
   );
+};
+
+MlNavigationTools.defaultProps = {
+  mapId: undefined,
+  show3DButton: true,
+  showFollowGpsButton: true,
+  showCenterLocationButton: false,
+  showZoomButtons: true,
+  sx: {
+    right: "5px",
+    bottom: "20px",
+  },
 };
 
 export default MlNavigationTools;